fix(docs): server-side error
